+++ internal/codegen/codegen.go
@@ -462,7 +462,10 @@ func NewHTTPClient(cfg *ClientConfig, scheme, hostname, port string) Client {
             TLSClientConfig: &tls.Config{
                 MinVersion: tlsMinVersion,
             },
-			ForceAttemptHTTP2: true,
+            ForceAttemptHTTP2:     true,
+            DisableCompression:    cfg.ClientDisableCompression,
+            ResponseHeaderTimeout: cfg.ClientResponseHeaderTimeout,
+            ExpectContinueTimeout: cfg.ClientExpectContinueTimeout,
         },
     }